vrrp: extract local-address resolution to instance_addr.go
Pure code-motion split of pkg/vrrp/instance.go (#5661). Move the local
IPv4/IPv6 and VIP-set address-resolution helpers out of the 2925-line
instance.go into a cohesive sibling file, package vrrp:

  interfaceAddrs, getLocalIP/setLocalIP, getLocalIPv6/setLocalIPv6,
  vipAddrSet, canonAddr, resolveLocalIPv4, reresolveLocalAddrs,
  resolveIPv6LinkLocal.

Declarations are moved verbatim — no rename of exported symbols, no
logic/timing/const change, and no reorder of any goroutine, defer,
channel, or mutex operation (whole functions moved intact). Imports
pruned by goimports; gofmt clean. Builds and `go test ./pkg/vrrp/...`
green.
